+++ client-connect/src/main/java/com/sails/client_connect/entity/User.java
@@ -59,11 +59,11 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int user_id;
+    private Long user_id;
     private String username;
     private String email;
     private String password;
-//    private String roles;
+
 
     @ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.PERSIST)
     @JoinTable(
